backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "a fastapi stater api"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "async-lru>=2.0.5",
    "beanie>=1.29.0",
    "cryptography>=45.0.1",
    "emails>=0.6",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "password-strength>=0.0.3.post2",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "uvicorn>=0.34.2",
    "dramatiq[redis,watch]>=1.17.0",
    "redis>=5.0.0",
    "questionary>=2.1.1",
    "python-dotenv>=1.0.0",
    "scalar-fastapi>=1.5.0",
]
